[MNG-5868] Adding serval times the same artifact via MavenProjectHelper (attachArtifact) does not produce a failure
o Updated method 'MavenProject.getAttachedArtifacts' to return an
  unmodifiable list.
o Updated method 'MavenProject.addAttachedArtifact' to throw a
  'DuplicateArtifactAttachmentException', if the same artifact already
  is attached.
o Updated all 'attachArtifact' methods of interface 'MavenProjectHelper'
  to also throw a 'DuplicateArtifactAttachmentException', if the
  same artifact already is attached.

This commit reduces the chances of 'hacking' the list of
attached artifacts in unspecified ways. Such 'hacks' will
now lead to 'RuntimeException's and will no longer work.
Depending on what this leads to, plugins need to be
updated and new methods may need to be added to
'MavenProjectHelper' for any use-cases no longer
supported. Reverting this commit to stay compatible
to unspecified behaviour should be the last option
considered.